+++ packages/frontend/web-ui/src/game/hooks/useGetGames/index.ts
@@ -1,4 +1,4 @@
-import { useSingleApiCall } from '../../../common/helpers/useSingleApiCall';
+import { useSingleApiCall } from '../../../common/hooks/useSingleApiCall';
 import { buildErrorMessage } from './utils/buildErrorMessage';
 import { buildRequestParams } from './utils/buildRequestParams';
 import { buildResult } from './utils/buildResult';
