+++ packages/frontend/web-ui/src/game/hooks/useGetGames/utils/buildErrorMessage.ts
