+++ packages/frontend/web-ui/src/game/hooks/useGetGames/utils/buildResult.spec.ts
